pyproject.toml
[project]
name = "pride_mcp_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mcp[cli]>=1.9.4",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.34.3",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]
